phishguard/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
dnspython==2.7.0
python-Levenshtein==0.26.1
confusables==1.2.0
reportlab==4.3.1
python-dotenv==1.1.0
